% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_check.R
\name{euclidean_check}
\alias{euclidean_check}
\title{Check Euclidean property of a Gram matrix}
\usage{
euclidean_check(G)
}
\arguments{
\item{G}{A symmetric Gram matrix (numeric matrix).}
}
\value{
A list with the following components:
\describe{
  \item{FNI}{Fraction of negative inertia, calculated as the absolute
    sum of negative eigenvalues divided by the absolute sum of all
    eigenvalues.}
  \item{is.Euclidean}{Indicator of Euclidean embeddability
    (1 = Euclidean, 0 = non-Euclidean).}
}
}
\description{
This function determines whether a given Gram (inner-product) matrix can be
embedded in Euclidean space and computes its fraction of negative inertia
(FNI).
}
