% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wunifrac.R
\name{wunifrac}
\alias{wunifrac}
\title{Compute Weighted UniFrac distance matrix from a phyloseq object (metric version)}
\usage{
wunifrac(physeq)
}
\arguments{
\item{physeq}{A \code{phyloseq} object with OTU table and rooted phylogenetic tree.}
}
\value{
A \code{dist} object representing the weighted UniFrac distance matrix between samples.
}
\description{
This function computes the weighted UniFrac distance between samples using the phylogenetic tree 
and normalized abundances stored in a \code{phyloseq} object.
}
\details{
This function manually computes the weighted UniFrac distances without using 
\code{phyloseq::UniFrac}. It supports custom workflows or further integration in pipelines.
}
